project save and new fixed

+++ src/python/image_graph.py
@@ -7,6 +7,14 @@ import os
 import shutil
 import cv2
 
+def get_pre_name(file):
+  pos = file.rfind('.')
+  return file[0:pos] if (pos > 0) else file
+
+def get_suffix(file):
+  pos = file.rfind('.')
+  return file[pos:] if (pos > 0) else '.json'
+
 def queue_nodes(g,nodes,node,func):
   for s in g.successors(node):
     queue_nodes(g,nodes,s,func)
@@ -23,15 +31,16 @@ class ImageGraph:
   def getUIGraph(self):
     return self.G
 
-  def set_name(self, name):
-    self.G.name = name
-
   def get_name(self):
     return self.G.name
 
-  def __init__(self, name, dir):
+  def __init__(self, pathname):
+    fname = os.path.split(pathname)[1]
+    name = get_pre_name(fname)
+    self.dir = os.path.abspath(os.path.split(pathname)[0])
     self.G = nx.DiGraph(name=name)
-    self.dir = os.path.abspath(dir)
+    if (os.path.exists(pathname)):
+      self.load(pathname)
 
   def get_nodes(self):
     return self.G.nodes()
@@ -39,21 +48,27 @@ class ImageGraph:
   def edges_iter(self, node):
     return self.G.edges_iter(node)
 
+  def get_pathname(self, name):
+     return os.path.join(self.dir, self.G.node[name]['file'])
+
   def get_edges(self):
     return self.G.edges()
 
-  def add_node(self,pathname, seriesname=None):
+  def add_node(self,pathname, seriesname=None, image=None):
     fname = os.path.split(pathname)[1]
-    origname = nname = fname[0:fname.rfind('.')]
-    suffix = fname[fname.rfind('.'):]
+    origname = nname = get_pre_name(fname)
+    suffix = get_suffix(fname)
     while (self.G.has_node(nname)):
       nname = nname + '_' + str(self.nextId())
       fname = nname + suffix
     newpathname = os.path.join(self.dir,fname)
     includePathInUndo = False
     if (not os.path.exists(newpathname)):
-      shutil.copy2(pathname, newpathname)
       includePathInUndo = True
+      if (os.path.exists(pathname)):
+        shutil.copy2(pathname, newpathname)
+      elif image is not None:
+        image.save(newpathname)
     self.G.add_node(nname, seriesname=(origname if seriesname is None else seriesname), file=fname, ownership=('yes' if includePathInUndo else 'no'))
     self.U = nx.DiGraph(name="undo")
     self.U.add_node(nname, action='addNode', file=fname, ownership=('yes' if includePathInUndo else 'no'))
@@ -103,13 +118,24 @@ class ImageGraph:
        im.load()
        return im
 
-  def remove(self,node,edgeFunc):
+  def remove(self,node,edgeFunc=None):
+    def maskRemover(edge):
+       if edgeFunc is not None:
+         edgeFunc(edge)
+       f = os.path.abspath(os.path.join(self.dir,edge['maskname']))
+       if (os.path.exists(f)):
+          os.remove(f)
     self.U = self.G.copy()
     for p in self.G.predecessors(node):
-      edgeFunc(self.G.edge[p][node])
-    nodes_to_remove = queue_nodes(self.G,[node],node,edgeFunc)
+      if edgeFunc is not None:
+        edgeFunc(self.G.edge[p][node])
+      maskRemover(self.G.edge[p][node])
+    nodes_to_remove = queue_nodes(self.G,[node],node,maskRemover)
     for n in nodes_to_remove:
       if (self.G.has_node(n)):
+        f = os.path.abspath(os.path.join(self.dir,self.G.node[n]['file']))
+        if (self.G.node[n]['ownership']=='yes' and os.path.exists(f)):
+          os.remove(f)
         self.G.remove_node(n)
 
   def predecessors(self,node):
@@ -124,14 +150,29 @@ class ImageGraph:
   def get_node(self,name):
     return self.G.node[name]
 
-  def load(self,fname):
-    with open(fname) as f:
+  def load(self,pathname):
+    with open(pathname) as f:
       d = json.load(f)
       self.G = json_graph.node_link_graph(d)
     if (self.G.has_node('idcount')):
       self.idc = self.G.node['idcount']['count']
       self.G.remove_node('idcount')
+    self.dir = os.path.abspath(os.path.split(pathname)[0])
      
+
+  def saveas(self, pathname):
+     currentdir = self.dir
+     fname = os.path.split(pathname)[1]
+     name = get_pre_name(fname)
+     self.dir = os.path.abspath(os.path.split(pathname)[0])
+     self.G.name = name
+     filename=os.path.abspath(os.path.join(self.dir,self.G.name + '.json'))
+     self.G.add_node('idcount',count=self.idc)
+     with open(filename, 'w') as f:
+        jg = json.dump(json_graph.node_link_data(self.G),f,indent=2)
+     self.G.remove_node('idcount')
+     self._copy_contents(currentdir)
+
   def save(self):
      filename=os.path.abspath(os.path.join(self.dir,self.G.name + '.json'))
      self.G.add_node('idcount',count=self.idc)
@@ -142,3 +183,17 @@ class ImageGraph:
   def nextId(self):
     self.idc+=1
     return self.idc
+
+  def _copy_contents(self, currentdir):
+    for nname in self.G.nodes():
+      node = self.G.node[nname]
+      oldpathname = os.path.join(currentdir,node['file'])
+      newpathname = os.path.join(self.dir,node['file'])
+      if (not os.path.exists(newpathname)):
+          shutil.copy2(oldpathname, newpathname)
+    for edgename in self.G.edges():
+      edge= self.G[edgename[0]][edgename[1]]
+      oldpathname = os.path.join(currentdir,edge['maskname'])
+      newpathname = os.path.join(self.dir,edge['maskname'])
+      if (not os.path.exists(newpathname)):
+          shutil.copy2(oldpathname, newpathname)
